+++ CHANGELOG.md
@@ -7,8 +7,9 @@
 - add type traits for resources
 - add some type-safe(should be) heap/buffer types
 - `ResourceDesc::buffer` now doesn't take an `alignment` argument
-- remove `CommittedResource`
-- `Device::create_committed_resource` now returns a `RawResource`
+- remove `CommittedResource`, `PlacedResource`
+- `Device::create_committed/placed_resource` now returns a `RawResource`
+- add methods for `PlacedBuffer` initialization
 
 # 0.2.1
 - Add a trait `factory::HwndProvider`, to bridge a hwnd provider with the swapchain creation API.